+++ docs/changelog.rst
@@ -11,6 +11,11 @@ Added
 
 - Support for property-level examples. `#467`_
 
+Fixed
+~~~~~
+
+- Content type conformance check for cases when Open API 3.0 schemas contain "default" response definitions. `#641`_
+
 `2.0.0`_ - 2020-07-01
 ---------------------
 
@@ -1209,6 +1214,7 @@ Fixed
 .. _0.3.0: https://github.com/kiwicom/schemathesis/compare/v0.2.0...v0.3.0
 .. _0.2.0: https://github.com/kiwicom/schemathesis/compare/v0.1.0...v0.2.0
 
+.. _#641: https://github.com/kiwicom/schemathesis/issues/641
 .. _#631: https://github.com/kiwicom/schemathesis/issues/631
 .. _#629: https://github.com/kiwicom/schemathesis/issues/629
 .. _#621: https://github.com/kiwicom/schemathesis/issues/621